split snow and sun simul
