+++ modelisation/snow/requirements.txt
@@ -1,0 +1,1 @@
+tartes~=1.3.3
