+++ modelisation/sun/requirements.txt
@@ -1,0 +1,2 @@
+scipy~=1.5.3
+solarpy~=0.1.3
