+++ requirements.txt
@@ -1,5 +1,2 @@
 numpy~=1.19.2
-matplotlib~=3.3.2
-scipy~=1.5.3
-solarpy~=0.1.3
-tartes~=1.3.3
+matplotlib==3.3.2
